refactor and add kagenti-a2a docs and scripts
