+++ src/client/components/DeployForm.tsx
@@ -1,189 +1,39 @@
 import React, { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { validateAgentName } from "../../shared/validate-agent-name.js";
-
-type InferenceProvider = "anthropic" | "openai" | "vertex-anthropic" | "vertex-google" | "custom-endpoint";
-type SecretRefSource = "env" | "file" | "exec";
-
-interface SecretRefValue {
-  source: SecretRefSource;
-  provider: string;
-  id: string;
-}
-
-interface DeployerInfo {
-  mode: string;
-  title: string;
-  description: string;
-  available: boolean;
-  unavailableReason?: string;
-  priority: number;
-  builtIn: boolean;
-  enabled: boolean;
-}
-
-interface Props {
-  onDeployStarted: (deployId: string) => void;
-}
-
-interface ServerDefaults {
-  hasAnthropicKey: boolean;
-  hasOpenaiKey: boolean;
-  hasTelegramToken: boolean;
-  telegramAllowFrom: string;
-  modelEndpoint: string;
-  prefix: string;
-  image: string;
-  containerRuntime?: string;
-  k8sAvailable?: boolean;
-  k8sContext?: string;
-  k8sNamespace?: string;
-  isOpenShift?: boolean;
-}
-
-interface GcpDefaults {
-  projectId: string | null;
-  location: string | null;
-  hasServiceAccountJson: boolean;
-  credentialType: string | null;
-  sources: {
-    projectId?: string;
-    location?: string;
-    credentials?: string;
-  };
-}
-
-interface SavedConfig {
-  name: string;
-  type: "local" | "k8s";
-  vars: Record<string, unknown>;
-}
-
-interface ModelEndpointOption {
-  id: string;
-  name: string;
-}
-
-const MODE_ICONS: Record<string, string> = {
-  local: "💻",
-  kubernetes: "☸️",
-  openshift: "☸️",
-  ssh: "🖥️",
-};
-
-const PROVIDER_OPTIONS: Array<{ id: InferenceProvider; label: string; desc: string }> = [
-  { id: "anthropic", label: "Anthropic", desc: "Claude models via Anthropic API" },
-  { id: "openai", label: "OpenAI", desc: "GPT models via OpenAI API" },
-  { id: "vertex-anthropic", label: "Google Vertex AI (Claude)", desc: "Claude models via Google Cloud" },
-  { id: "vertex-google", label: "Google Vertex AI (Gemini)", desc: "Gemini models via Google Cloud" },
-  { id: "custom-endpoint", label: "Model Endpoint", desc: "OpenAI-compatible self-hosted model server" },
-];
-
-const MODEL_DEFAULTS: Record<InferenceProvider, string> = {
-  "anthropic": "claude-sonnet-4-6",
-  "openai": "openai/gpt-5",
-  "vertex-anthropic": "anthropic-vertex/claude-sonnet-4-6",
-  "vertex-google": "google-vertex/gemini-2.5-pro",
-  "custom-endpoint": "",
-};
-
-const MODEL_HINTS: Record<InferenceProvider, string> = {
-  "anthropic": "Examples: claude-sonnet-4-6, claude-opus-4-6, claude-haiku-4-5",
-  "openai": "Examples: openai/gpt-5, openai/gpt-5.3",
-  "vertex-anthropic": "Examples: anthropic-vertex/claude-sonnet-4-6, anthropic-vertex/claude-opus-4-6",
-  "vertex-google": "Examples: google-vertex/gemini-2.5-pro, google-vertex/gemini-2.5-flash",
-  "custom-endpoint": "Examples: mistral-small-24b-w8a8, mistral-medium-3.1-24b-instruct-2506",
-};
-
-const PROXY_MODEL_HINTS: Record<string, string> = {
-  "vertex-anthropic": "Examples: claude-sonnet-4-6, claude-opus-4-6, claude-haiku-4-5",
-  "vertex-google": "Examples: gemini-2.5-pro, gemini-2.5-flash",
-};
-
-function defaultImageForProvider(provider: InferenceProvider): string {
-  return provider === "vertex-anthropic"
-    ? "ghcr.io/openclaw/openclaw:latest"
-    : "ghcr.io/openclaw/openclaw:latest";
-}
-
-function decodeBase64(value: string | undefined): string {
-  if (!value) return "";
-  try {
-    return window.atob(value);
-  } catch {
-    return "";
-  }
-}
-
-function encodeBase64(value: string): string {
-  return window.btoa(value);
-}
-
-function decodeJsonBase64<T>(value: string | undefined): T | undefined {
-  const decoded = decodeBase64(value);
-  if (!decoded) return undefined;
-  try {
-    return JSON.parse(decoded) as T;
-  } catch {
-    return undefined;
-  }
-}
-
-function trimToUndefined(value: string): string | undefined {
-  const trimmed = value.trim();
-  return trimmed ? trimmed : undefined;
-}
-
-function buildSecretRef(source: string, provider: string, id: string): SecretRefValue | undefined {
-  const trimmedProvider = provider.trim();
-  const trimmedId = id.trim();
-  if (!source && !trimmedProvider && !trimmedId) return undefined;
-  if ((source !== "env" && source !== "file" && source !== "exec") || !trimmedProvider || !trimmedId) {
-    return undefined;
-  }
-  return {
-    source,
-    provider: trimmedProvider,
-    id: trimmedId,
-  } as SecretRefValue;
-}
-
-function inferAgentNameFromPath(value: string): string {
-  const trimmed = value.trim().replace(/[\\/]+$/, "");
-  if (!trimmed) return "";
-  const parts = trimmed.split(/[\\/]/).filter(Boolean);
-  const base = parts[parts.length - 1] || "";
-  return base
-    .toLowerCase()
-    .replace(/[^a-z0-9._-]+/g, "-")
-    .replace(/^-+|-+$/g, "");
-}
-
-function inferDisplayNameFromAgentName(value: string): string {
-  return value
-    .split(/[-_.]+/)
-    .filter(Boolean)
-    .map((part) => part.charAt(0).toUpperCase() + part.slice(1))
-    .join(" ");
-}
-
-function sanitizeNamespacePart(value: string): string {
-  return value
-    .trim()
-    .toLowerCase()
-    .replace(/[^a-z0-9-]+/g, "-")
-    .replace(/-+/g, "-")
-    .replace(/^-+|-+$/g, "");
-}
-
-function deriveNamespace(prefix: string, agentName: string): string {
-  const cleanPrefix = sanitizeNamespacePart(prefix) || "user";
-  const cleanAgent = sanitizeNamespacePart(agentName) || "agent";
-  return `${cleanPrefix}-${cleanAgent}-openclaw`;
-}
+import {
+  MODE_ICONS,
+} from "./deploy-form/constants.js";
+import {
+  defaultImageForProvider,
+  deriveNamespace,
+  inferAgentNameFromPath,
+  inferDisplayNameFromAgentName,
+  buildSecretRef,
+} from "./deploy-form/utils.js";
+import {
+  applySavedVarsToConfig,
+  buildDeployRequestBody,
+  buildEnvFileContent,
+  createInitialDeployFormConfig,
+  inferSavedInferenceProvider,
+} from "./deploy-form/serialization.js";
+import { ProviderSection } from "./deploy-form/ProviderSection.js";
+import { SandboxSection } from "./deploy-form/SandboxSection.js";
+import { SecretProvidersSection } from "./deploy-form/SecretProvidersSection.js";
+import type {
+  DeployFormConfig,
+  DeployerInfo,
+  DeployFormProps,
+  GcpDefaults,
+  InferenceProvider,
+  ModelEndpointOption,
+  SavedConfig,
+  ServerDefaults,
+} from "./deploy-form/types.js";
 
 const LAST_AGENT_SOURCE_DIR_KEY = "openclaw:last-agent-source-dir";
 
-export default function DeployForm({ onDeployStarted }: Props) {
+export default function DeployForm({ onDeployStarted }: DeployFormProps) {
   const [mode, setMode] = useState("local");
   const [deploying, setDeploying] = useState(false);
   const [defaults, setDefaults] = useState<ServerDefaults | null>(null);
@@ -192,85 +42,7 @@ export default function DeployForm({ onDeployStarted }: Props) {
   const [loadedConfigLabel, setLoadedConfigLabel] = useState<string | null>(null);
   const [autoLoadedEnvDir, setAutoLoadedEnvDir] = useState<string | null>(null);
   const [inferenceProvider, setInferenceProvider] = useState<InferenceProvider>("anthropic");
-  const [config, setConfig] = useState({
-    prefix: "",
-    agentName: "",
-    agentDisplayName: "",
-    image: "",
-    containerRunArgs: "",
-    secretsProvidersJson: "",
-    anthropicApiKeyRefSource: "env",
-    anthropicApiKeyRefProvider: "default",
-    anthropicApiKeyRefId: "",
-    openaiApiKeyRefSource: "env",
-    openaiApiKeyRefProvider: "default",
-    openaiApiKeyRefId: "",
-    telegramBotTokenRefSource: "env",
-    telegramBotTokenRefProvider: "default",
-    telegramBotTokenRefId: "",
-    sandboxEnabled: false,
-    sandboxMode: "all",
-    sandboxScope: "session",
-    sandboxWorkspaceAccess: "rw",
-    sandboxToolPolicyEnabled: false,
-    sandboxToolAllowFiles: true,
-    sandboxToolAllowSessions: true,
-    sandboxToolAllowMemory: true,
-    sandboxToolAllowRuntime: false,
-    sandboxToolAllowBrowser: false,
-    sandboxToolAllowAutomation: false,
-    sandboxToolAllowMessaging: false,
-    sandboxSshTarget: "",
-    sandboxSshWorkspaceRoot: "/tmp/openclaw-sandboxes",
-    sandboxSshStrictHostKeyChecking: true,
-    sandboxSshUpdateHostKeys: true,
-    sandboxSshIdentityPath: "",
-    sandboxSshCertificate: "",
-    sandboxSshCertificatePath: "",
-    sandboxSshKnownHosts: "",
-    sandboxSshKnownHostsPath: "",
-    anthropicApiKey: "",
-    openaiApiKey: "",
-    anthropicModel: "",
-    openaiModel: "",
-    agentModel: "",
-    openaiCompatibleEndpointsEnabled: true,
-    modelEndpoint: "",
-    modelEndpointApiKey: "",
-    modelEndpointModel: "",
-    modelEndpointModelLabel: "",
-    modelEndpointModels: [] as ModelEndpointOption[],
-    port: "18789",
-    // Vertex AI / GCP
-    googleCloudProject: "",
-    googleCloudLocation: "",
-    gcpServiceAccountJson: "",
-    gcpServiceAccountPath: "",
-    // SSH fields
-    sshHost: "",
-    sshUser: "",
-    // Agent provisioning
-    agentSourceDir: "",
-    // Telegram
-    telegramEnabled: false,
-    telegramBotToken: "",
-    telegramAllowFrom: "",
-    // Agent security
-    cronEnabled: false,
-    subagentPolicy: "none" as "none" | "self" | "unrestricted",
-    // Kubernetes
-    namespace: "",
-    withA2a: false,
-    a2aRealm: "",
-    a2aKeycloakNamespace: "keycloak",
-    // LiteLLM proxy
-    litellmProxy: true,
-    // OTEL tracing
-    otelEnabled: false,
-    otelJaeger: false,
-    otelEndpoint: "",
-    otelExperimentId: "",
-  });
+  const [config, setConfig] = useState<DeployFormConfig>(createInitialDeployFormConfig);
 
   const [gcpDefaults, setGcpDefaults] = useState<GcpDefaults | null>(null);
   const [gcpDefaultsFetched, setGcpDefaultsFetched] = useState(false);
@@ -463,190 +235,13 @@ export default function DeployForm({ onDeployStarted }: Props) {
   }, [config.modelEndpointModels]);
 
   const applyVars = (vars: Record<string, unknown>) => {
-    // Support both local .env keys (OPENCLAW_PREFIX) and K8s JSON keys (prefix)
-    const v = (envKey: string, jsonKey: string) => {
-      const value = vars[envKey] ?? vars[jsonKey];
-      return typeof value === "string" ? value : "";
-    };
-    const anthropicApiKeyRef =
-      decodeJsonBase64<SecretRefValue>(vars.ANTHROPIC_API_KEY_REF_B64)
-      || (typeof (vars as Record<string, unknown>).anthropicApiKeyRef === "object"
-        ? (vars as unknown as { anthropicApiKeyRef?: SecretRefValue }).anthropicApiKeyRef
-        : undefined);
-    const openaiApiKeyRef =
-      decodeJsonBase64<SecretRefValue>(vars.OPENAI_API_KEY_REF_B64)
-      || (typeof (vars as Record<string, unknown>).openaiApiKeyRef === "object"
-        ? (vars as unknown as { openaiApiKeyRef?: SecretRefValue }).openaiApiKeyRef
-        : undefined);
-    const telegramBotTokenRef =
-      decodeJsonBase64<SecretRefValue>(vars.TELEGRAM_BOT_TOKEN_REF_B64)
-      || (typeof (vars as Record<string, unknown>).telegramBotTokenRef === "object"
-        ? (vars as unknown as { telegramBotTokenRef?: SecretRefValue }).telegramBotTokenRef
-        : undefined);
-    const savedProvidersJson =
-      decodeBase64(vars.SECRETS_PROVIDERS_JSON_B64)
-      || (typeof (vars as Record<string, unknown>).secretsProvidersJson === "string"
-        ? (vars as unknown as { secretsProvidersJson?: string }).secretsProvidersJson
-        : "");
-    const explicitNamespace = v("K8S_NAMESPACE", "namespace");
-    const savedEndpointModels =
-      decodeJsonBase64<ModelEndpointOption[]>(vars.MODEL_ENDPOINT_MODELS_B64)
-      || (Array.isArray((vars as Record<string, unknown>).modelEndpointModels)
-        ? ((vars as Record<string, unknown>).modelEndpointModels as ModelEndpointOption[])
-        : undefined);
-    const savedInferenceProvider = v("INFERENCE_PROVIDER", "inferenceProvider");
-    if (
-      savedInferenceProvider === "anthropic"
-      || savedInferenceProvider === "openai"
-      || savedInferenceProvider === "vertex-anthropic"
-      || savedInferenceProvider === "vertex-google"
-      || savedInferenceProvider === "custom-endpoint"
-    ) {
-      setInferenceProvider(savedInferenceProvider);
-    } else {
-      const vertexEnabled = vars.VERTEX_ENABLED === "true" || vars.vertexEnabled === "true";
-      if (vertexEnabled) {
-        const vp = vars.VERTEX_PROVIDER || vars.vertexProvider || "anthropic";
-        setInferenceProvider(vp === "google" ? "vertex-google" : "vertex-anthropic");
-      } else if (v("MODEL_ENDPOINT", "modelEndpoint") || openaiApiKeyRef) {
-        setInferenceProvider("custom-endpoint");
-      } else if (v("ANTHROPIC_API_KEY", "anthropicApiKey") || anthropicApiKeyRef) {
-        setInferenceProvider("anthropic");
-      } else if (v("OPENAI_API_KEY", "openaiApiKey") || openaiApiKeyRef) {
-        setInferenceProvider("openai");
-      }
+    const nextInferenceProvider = inferSavedInferenceProvider(vars);
+    if (nextInferenceProvider) {
+      setInferenceProvider(nextInferenceProvider);
     }
-
-    setNamespaceManuallyEdited(Boolean(explicitNamespace));
-
-    setConfig((prev) => ({
-      ...prev,
-      prefix: v("OPENCLAW_PREFIX", "prefix") || prev.prefix,
-      agentName: v("OPENCLAW_AGENT_NAME", "agentName") || prev.agentName,
-      agentDisplayName: v("OPENCLAW_DISPLAY_NAME", "agentDisplayName") || prev.agentDisplayName,
-      image: v("OPENCLAW_IMAGE", "image") || prev.image,
-      containerRunArgs: v("OPENCLAW_CONTAINER_RUN_ARGS", "containerRunArgs") || prev.containerRunArgs,
-      secretsProvidersJson: savedProvidersJson || prev.secretsProvidersJson,
-      anthropicApiKeyRefSource: anthropicApiKeyRef?.source || prev.anthropicApiKeyRefSource,
-      anthropicApiKeyRefProvider: anthropicApiKeyRef?.provider || prev.anthropicApiKeyRefProvider,
-      anthropicApiKeyRefId: anthropicApiKeyRef?.id || prev.anthropicApiKeyRefId,
-      openaiApiKeyRefSource: openaiApiKeyRef?.source || prev.openaiApiKeyRefSource,
-      openaiApiKeyRefProvider: openaiApiKeyRef?.provider || prev.openaiApiKeyRefProvider,
-      openaiApiKeyRefId: openaiApiKeyRef?.id || prev.openaiApiKeyRefId,
-      telegramBotTokenRefSource: telegramBotTokenRef?.source || prev.telegramBotTokenRefSource,
-      telegramBotTokenRefProvider: telegramBotTokenRef?.provider || prev.telegramBotTokenRefProvider,
-      telegramBotTokenRefId: telegramBotTokenRef?.id || prev.telegramBotTokenRefId,
-      sandboxEnabled:
-        vars.SANDBOX_ENABLED === "true" || vars.sandboxEnabled === "true" || prev.sandboxEnabled,
-      sandboxMode: v("SANDBOX_MODE", "sandboxMode") || prev.sandboxMode,
-      sandboxScope: v("SANDBOX_SCOPE", "sandboxScope") || prev.sandboxScope,
-      sandboxToolPolicyEnabled:
-        vars.SANDBOX_TOOL_POLICY_ENABLED === "true"
-          || vars.sandboxToolPolicyEnabled === "true"
-          || prev.sandboxToolPolicyEnabled,
-      sandboxToolAllowFiles:
-        vars.SANDBOX_TOOL_ALLOW_FILES === "false"
-          ? false
-          : vars.sandboxToolAllowFiles === "false"
-            ? false
-            : prev.sandboxToolAllowFiles,
-      sandboxToolAllowSessions:
-        vars.SANDBOX_TOOL_ALLOW_SESSIONS === "false"
-          ? false
-          : vars.sandboxToolAllowSessions === "false"
-            ? false
-            : prev.sandboxToolAllowSessions,
-      sandboxToolAllowMemory:
-        vars.SANDBOX_TOOL_ALLOW_MEMORY === "false"
-          ? false
-          : vars.sandboxToolAllowMemory === "false"
-            ? false
-            : prev.sandboxToolAllowMemory,
-      sandboxToolAllowRuntime:
-        vars.SANDBOX_TOOL_ALLOW_RUNTIME === "true"
-          || vars.sandboxToolAllowRuntime === "true"
-          || prev.sandboxToolAllowRuntime,
-      sandboxToolAllowBrowser:
-        vars.SANDBOX_TOOL_ALLOW_BROWSER === "true"
-          || vars.sandboxToolAllowBrowser === "true"
-          || prev.sandboxToolAllowBrowser,
-      sandboxToolAllowAutomation:
-        vars.SANDBOX_TOOL_ALLOW_AUTOMATION === "true"
-          || vars.sandboxToolAllowAutomation === "true"
-          || prev.sandboxToolAllowAutomation,
-      sandboxToolAllowMessaging:
-        vars.SANDBOX_TOOL_ALLOW_MESSAGING === "true"
-          || vars.sandboxToolAllowMessaging === "true"
-          || prev.sandboxToolAllowMessaging,
-      sandboxWorkspaceAccess:
-        v("SANDBOX_WORKSPACE_ACCESS", "sandboxWorkspaceAccess") || prev.sandboxWorkspaceAccess,
-      sandboxSshTarget:
-        v("SANDBOX_SSH_TARGET", "sandboxSshTarget") || prev.sandboxSshTarget,
-      sandboxSshWorkspaceRoot:
-        v("SANDBOX_SSH_WORKSPACE_ROOT", "sandboxSshWorkspaceRoot") ||
-        prev.sandboxSshWorkspaceRoot,
-      sandboxSshIdentityPath:
-        v("SANDBOX_SSH_IDENTITY_PATH", "sandboxSshIdentityPath") || prev.sandboxSshIdentityPath,
-      sandboxSshCertificatePath:
-        v("SANDBOX_SSH_CERTIFICATE_PATH", "sandboxSshCertificatePath") ||
-        prev.sandboxSshCertificatePath,
-      sandboxSshKnownHostsPath:
-        v("SANDBOX_SSH_KNOWN_HOSTS_PATH", "sandboxSshKnownHostsPath") ||
-        prev.sandboxSshKnownHostsPath,
-      sandboxSshStrictHostKeyChecking:
-        vars.SANDBOX_SSH_STRICT_HOST_KEY_CHECKING === "false"
-          ? false
-          : vars.sandboxSshStrictHostKeyChecking === "false"
-            ? false
-            : prev.sandboxSshStrictHostKeyChecking,
-      sandboxSshUpdateHostKeys:
-        vars.SANDBOX_SSH_UPDATE_HOST_KEYS === "false"
-          ? false
-          : vars.sandboxSshUpdateHostKeys === "false"
-            ? false
-            : prev.sandboxSshUpdateHostKeys,
-      sandboxSshCertificate:
-        decodeBase64(vars.SANDBOX_SSH_CERTIFICATE_B64) ||
-        v("sandboxSshCertificate", "sandboxSshCertificate") ||
-        prev.sandboxSshCertificate,
-      sandboxSshKnownHosts:
-        decodeBase64(vars.SANDBOX_SSH_KNOWN_HOSTS_B64) ||
-        v("sandboxSshKnownHosts", "sandboxSshKnownHosts") ||
-        prev.sandboxSshKnownHosts,
-      port: v("OPENCLAW_PORT", "port") || prev.port,
-      anthropicModel: v("ANTHROPIC_MODEL", "anthropicModel") || prev.anthropicModel,
-      openaiModel: v("OPENAI_MODEL", "openaiModel") || prev.openaiModel,
-      agentModel: v("AGENT_MODEL", "agentModel") || prev.agentModel,
-      openaiCompatibleEndpointsEnabled:
-        vars.OPENAI_COMPATIBLE_ENDPOINTS_ENABLED === "false"
-          ? false
-          : vars.openaiCompatibleEndpointsEnabled === false
-            ? false
-            : prev.openaiCompatibleEndpointsEnabled,
-      modelEndpoint: v("MODEL_ENDPOINT", "modelEndpoint") || prev.modelEndpoint,
-      modelEndpointApiKey: v("MODEL_ENDPOINT_API_KEY", "modelEndpointApiKey") || prev.modelEndpointApiKey,
-      modelEndpointModel: v("MODEL_ENDPOINT_MODEL", "modelEndpointModel") || prev.modelEndpointModel,
-      modelEndpointModelLabel: v("MODEL_ENDPOINT_MODEL_LABEL", "modelEndpointModelLabel") || prev.modelEndpointModelLabel,
-      modelEndpointModels: savedEndpointModels || prev.modelEndpointModels,
-      googleCloudProject: v("GOOGLE_CLOUD_PROJECT", "googleCloudProject") || prev.googleCloudProject,
-      googleCloudLocation: v("GOOGLE_CLOUD_LOCATION", "googleCloudLocation") || prev.googleCloudLocation,
-      agentSourceDir: v("AGENT_SOURCE_DIR", "agentSourceDir") || prev.agentSourceDir,
-      telegramBotToken: v("TELEGRAM_BOT_TOKEN", "telegramBotToken") || prev.telegramBotToken,
-      telegramAllowFrom: v("TELEGRAM_ALLOW_FROM", "telegramAllowFrom") || prev.telegramAllowFrom,
-      namespace: explicitNamespace || prev.namespace,
-      withA2a: vars.WITH_A2A === "true" || vars.withA2a === "true" || prev.withA2a,
-      a2aRealm: v("A2A_REALM", "a2aRealm") || prev.a2aRealm,
-      a2aKeycloakNamespace:
-        v("A2A_KEYCLOAK_NAMESPACE", "a2aKeycloakNamespace") || prev.a2aKeycloakNamespace,
-      litellmProxy: vars.litellmProxy === "false" ? false : prev.litellmProxy,
-      otelEnabled: vars.OTEL_ENABLED === "true" || vars.otelEnabled === "true" || prev.otelEnabled,
-      otelJaeger: vars.OTEL_JAEGER === "true" || vars.otelJaeger === "true" || prev.otelJaeger,
-      otelEndpoint: v("OTEL_ENDPOINT", "otelEndpoint") || prev.otelEndpoint,
-      otelExperimentId: v("OTEL_EXPERIMENT_ID", "otelExperimentId") || prev.otelExperimentId,
-      otelImage: v("OTEL_IMAGE", "otelImage") || prev.otelImage,
-      cronEnabled: vars.cronEnabled === "true" ? true : prev.cronEnabled,
-      subagentPolicy: (vars.subagentPolicy as "none" | "self" | "unrestricted") || prev.subagentPolicy,
-    }));
+    const applied = applySavedVarsToConfig(vars, config);
+    setNamespaceManuallyEdited(applied.namespaceManuallyEdited);
+    setConfig(applied.config);
   };
 
   const [displayNameManuallyEdited, setDisplayNameManuallyEdited] = useState(false);
@@ -782,98 +377,16 @@ export default function DeployForm({ onDeployStarted }: Props) {
     }
     setDeploying(true);
     try {
-      const vertexEnabled = isVertex;
-      const vertexProvider = inferenceProvider === "vertex-google" ? "google" : "anthropic";
-
-      const body = {
+      const body = buildDeployRequestBody({
         mode,
         inferenceProvider,
-        prefix: config.prefix,
-        agentName: config.agentName,
-        agentDisplayName: config.agentDisplayName || config.agentName,
-        image: trimToUndefined(config.image),
-        containerRunArgs: mode === "local" ? trimToUndefined(config.containerRunArgs) : undefined,
-        secretsProvidersJson: trimToUndefined(config.secretsProvidersJson),
+        config,
+        isVertex,
+        suggestedNamespace,
         anthropicApiKeyRef,
         openaiApiKeyRef,
-        telegramBotTokenRef:
-          config.telegramEnabled ? telegramBotTokenRef : undefined,
-        sandboxEnabled: config.sandboxEnabled || undefined,
-        sandboxBackend: config.sandboxEnabled ? "ssh" : undefined,
-        sandboxMode: config.sandboxEnabled ? config.sandboxMode : undefined,
-        sandboxScope: config.sandboxEnabled ? config.sandboxScope : undefined,
-        sandboxToolPolicyEnabled:
-          config.sandboxEnabled ? config.sandboxToolPolicyEnabled || undefined : undefined,
-        sandboxToolAllowFiles:
-          config.sandboxEnabled ? config.sandboxToolAllowFiles : undefined,
-        sandboxToolAllowSessions:
-          config.sandboxEnabled ? config.sandboxToolAllowSessions : undefined,
-        sandboxToolAllowMemory:
-          config.sandboxEnabled ? config.sandboxToolAllowMemory : undefined,
-        sandboxToolAllowRuntime:
-          config.sandboxEnabled ? config.sandboxToolAllowRuntime : undefined,
-        sandboxToolAllowBrowser:
-          config.sandboxEnabled ? config.sandboxToolAllowBrowser : undefined,
-        sandboxToolAllowAutomation:
-          config.sandboxEnabled ? config.sandboxToolAllowAutomation : undefined,
-        sandboxToolAllowMessaging:
-          config.sandboxEnabled ? config.sandboxToolAllowMessaging : undefined,
-        sandboxWorkspaceAccess: config.sandboxEnabled ? config.sandboxWorkspaceAccess : undefined,
-        sandboxSshTarget: config.sandboxEnabled ? config.sandboxSshTarget || undefined : undefined,
-        sandboxSshWorkspaceRoot:
-          config.sandboxEnabled ? config.sandboxSshWorkspaceRoot || undefined : undefined,
-        sandboxSshIdentityPath:
-          config.sandboxEnabled ? config.sandboxSshIdentityPath || undefined : undefined,
-        sandboxSshCertificatePath:
-          config.sandboxEnabled ? config.sandboxSshCertificatePath || undefined : undefined,
-        sandboxSshKnownHostsPath:
-          config.sandboxEnabled ? config.sandboxSshKnownHostsPath || undefined : undefined,
-        sandboxSshStrictHostKeyChecking:
-          config.sandboxEnabled ? config.sandboxSshStrictHostKeyChecking : undefined,
-        sandboxSshUpdateHostKeys:
-          config.sandboxEnabled ? config.sandboxSshUpdateHostKeys : undefined,
-        sandboxSshCertificate:
-          config.sandboxEnabled ? config.sandboxSshCertificate || undefined : undefined,
-        sandboxSshKnownHosts:
-          config.sandboxEnabled ? config.sandboxSshKnownHosts || undefined : undefined,
-        anthropicApiKey: !anthropicApiKeyRef ? trimToUndefined(config.anthropicApiKey) : undefined,
-        openaiApiKey: !openaiApiKeyRef ? trimToUndefined(config.openaiApiKey) : undefined,
-        anthropicModel: trimToUndefined(config.anthropicModel),
-        openaiModel: trimToUndefined(config.openaiModel),
-        agentModel: config.agentModel || undefined,
-        openaiCompatibleEndpointsEnabled: config.openaiCompatibleEndpointsEnabled,
-        modelEndpoint: trimToUndefined(config.modelEndpoint),
-        modelEndpointApiKey: trimToUndefined(config.modelEndpointApiKey),
-        modelEndpointModel: trimToUndefined(config.modelEndpointModel),
-        modelEndpointModelLabel: trimToUndefined(config.modelEndpointModelLabel),
-        modelEndpointModels: config.modelEndpointModels.length > 0 ? config.modelEndpointModels : undefined,
-        port: parseInt(config.port, 10) || 18789,
-        vertexEnabled: vertexEnabled || undefined,
-        vertexProvider: vertexEnabled ? vertexProvider : undefined,
-        googleCloudProject: vertexEnabled ? trimToUndefined(config.googleCloudProject) : undefined,
-        googleCloudLocation: vertexEnabled ? trimToUndefined(config.googleCloudLocation) : undefined,
-        gcpServiceAccountJson: vertexEnabled ? trimToUndefined(config.gcpServiceAccountJson) : undefined,
-        gcpServiceAccountPath: vertexEnabled ? trimToUndefined(config.gcpServiceAccountPath) : undefined,
-        litellmProxy: vertexEnabled ? config.litellmProxy : undefined,
-        namespace: trimToUndefined(config.namespace) || suggestedNamespace || undefined,
-        withA2a: config.withA2a || undefined,
-        a2aRealm: config.withA2a ? trimToUndefined(config.a2aRealm) : undefined,
-        a2aKeycloakNamespace:
-          config.withA2a ? trimToUndefined(config.a2aKeycloakNamespace) : undefined,
-        sshHost: trimToUndefined(config.sshHost),
-        sshUser: trimToUndefined(config.sshUser),
-        agentSourceDir: trimToUndefined(config.agentSourceDir),
-        telegramEnabled: config.telegramEnabled || undefined,
-        telegramBotToken:
-          config.telegramEnabled && !telegramBotTokenRef ? trimToUndefined(config.telegramBotToken) : undefined,
-        telegramAllowFrom: config.telegramEnabled ? trimToUndefined(config.telegramAllowFrom) : undefined,
-        otelEnabled: config.otelEnabled || undefined,
-        otelJaeger: config.otelEnabled ? config.otelJaeger || undefined : undefined,
-        otelEndpoint: config.otelEnabled ? trimToUndefined(config.otelEndpoint) : undefined,
-        otelExperimentId: config.otelEnabled ? trimToUndefined(config.otelExperimentId) : undefined,
-        cronEnabled: config.cronEnabled || undefined,
-        subagentPolicy: config.subagentPolicy !== "none" ? config.subagentPolicy : undefined,
-      };
+        telegramBotTokenRef,
+      });
 
       const res = await fetch("/api/deploy", {
         method: "POST",
@@ -893,91 +406,15 @@ export default function DeployForm({ onDeployStarted }: Props) {
   };
 
   const handleEnvDownload = () => {
-    const lines = [
-      "# OpenClaw installer config",
-      `OPENCLAW_PREFIX=${config.prefix}`,
-      `OPENCLAW_AGENT_NAME=${config.agentName}`,
-      `OPENCLAW_DISPLAY_NAME=${config.agentDisplayName}`,
-      `OPENCLAW_IMAGE=${config.image}`,
-      `OPENCLAW_CONTAINER_RUN_ARGS=${config.containerRunArgs}`,
-      `OPENCLAW_PORT=${config.port}`,
-      `AGENT_SOURCE_DIR=${config.agentSourceDir}`,
-      "",
-      `INFERENCE_PROVIDER=${inferenceProvider}`,
-      `ANTHROPIC_API_KEY=${anthropicApiKeyRef ? "" : config.anthropicApiKey}`,
-      `OPENAI_API_KEY=${openaiApiKeyRef ? "" : config.openaiApiKey}`,
-      `ANTHROPIC_MODEL=${config.anthropicModel}`,
-      `OPENAI_MODEL=${config.openaiModel}`,
-      `OPENAI_COMPATIBLE_ENDPOINTS_ENABLED=${config.openaiCompatibleEndpointsEnabled}`,
-      `MODEL_ENDPOINT=${config.modelEndpoint}`,
-      `MODEL_ENDPOINT_API_KEY=${config.modelEndpointApiKey}`,
-      `MODEL_ENDPOINT_MODEL=${config.modelEndpointModel}`,
-      `MODEL_ENDPOINT_MODEL_LABEL=${config.modelEndpointModelLabel}`,
-      `MODEL_ENDPOINT_MODELS_B64=${encodeBase64(JSON.stringify(config.modelEndpointModels))}`,
-      `AGENT_MODEL=${config.agentModel}`,
-      "",
-      `VERTEX_ENABLED=${isVertex}`,
-      `VERTEX_PROVIDER=${inferenceProvider === "vertex-google" ? "google" : "anthropic"}`,
-      `GOOGLE_CLOUD_PROJECT=${config.googleCloudProject}`,
-      `GOOGLE_CLOUD_LOCATION=${config.googleCloudLocation}`,
-      `GCP_SERVICE_ACCOUNT_PATH=${config.gcpServiceAccountPath}`,
-      `LITELLM_PROXY=${config.litellmProxy}`,
-      "",
-      `SANDBOX_ENABLED=${config.sandboxEnabled}`,
-      "SANDBOX_BACKEND=ssh",
-      `SANDBOX_MODE=${config.sandboxMode}`,
-      `SANDBOX_SCOPE=${config.sandboxScope}`,
-      `SANDBOX_WORKSPACE_ACCESS=${config.sandboxWorkspaceAccess}`,
-      `SANDBOX_SSH_TARGET=${config.sandboxSshTarget}`,
-      `SANDBOX_SSH_WORKSPACE_ROOT=${config.sandboxSshWorkspaceRoot}`,
-      `SANDBOX_SSH_IDENTITY_PATH=${config.sandboxSshIdentityPath}`,
-      `SANDBOX_SSH_CERTIFICATE_PATH=${config.sandboxSshCertificatePath}`,
-      `SANDBOX_SSH_KNOWN_HOSTS_PATH=${config.sandboxSshKnownHostsPath}`,
-      `SANDBOX_SSH_STRICT_HOST_KEY_CHECKING=${config.sandboxSshStrictHostKeyChecking}`,
-      `SANDBOX_SSH_UPDATE_HOST_KEYS=${config.sandboxSshUpdateHostKeys}`,
-      `SANDBOX_TOOL_POLICY_ENABLED=${config.sandboxToolPolicyEnabled}`,
-      `SANDBOX_TOOL_ALLOW_FILES=${config.sandboxToolAllowFiles}`,
-      `SANDBOX_TOOL_ALLOW_SESSIONS=${config.sandboxToolAllowSessions}`,
-      `SANDBOX_TOOL_ALLOW_MEMORY=${config.sandboxToolAllowMemory}`,
-      `SANDBOX_TOOL_ALLOW_RUNTIME=${config.sandboxToolAllowRuntime}`,
-      `SANDBOX_TOOL_ALLOW_BROWSER=${config.sandboxToolAllowBrowser}`,
-      `SANDBOX_TOOL_ALLOW_AUTOMATION=${config.sandboxToolAllowAutomation}`,
-      `SANDBOX_TOOL_ALLOW_MESSAGING=${config.sandboxToolAllowMessaging}`,
-      "",
-      `TELEGRAM_ENABLED=${config.telegramEnabled}`,
-      `TELEGRAM_BOT_TOKEN=${telegramBotTokenRef ? "" : config.telegramBotToken}`,
-      `TELEGRAM_ALLOW_FROM=${config.telegramAllowFrom}`,
-      `OTEL_ENABLED=${config.otelEnabled}`,
-      `OTEL_JAEGER=${config.otelJaeger}`,
-      `OTEL_ENDPOINT=${config.otelEndpoint}`,
-      `OTEL_EXPERIMENT_ID=${config.otelExperimentId}`,
-      "",
-      `K8S_NAMESPACE=${config.namespace || suggestedNamespace}`,
-      `WITH_A2A=${config.withA2a}`,
-      `A2A_REALM=${config.a2aRealm}`,
-      `A2A_KEYCLOAK_NAMESPACE=${config.a2aKeycloakNamespace}`,
-    ];
-
-    if (config.sandboxSshCertificate && !config.sandboxSshCertificatePath) {
-      lines.push(`SANDBOX_SSH_CERTIFICATE_B64=${encodeBase64(config.sandboxSshCertificate)}`);
-    }
-    if (config.sandboxSshKnownHosts && !config.sandboxSshKnownHostsPath) {
-      lines.push(`SANDBOX_SSH_KNOWN_HOSTS_B64=${encodeBase64(config.sandboxSshKnownHosts)}`);
-    }
-    if (config.secretsProvidersJson.trim()) {
-      lines.push(`SECRETS_PROVIDERS_JSON_B64=${encodeBase64(config.secretsProvidersJson)}`);
-    }
-    if (anthropicApiKeyRef) {
-      lines.push(`ANTHROPIC_API_KEY_REF_B64=${encodeBase64(JSON.stringify(anthropicApiKeyRef))}`);
-    }
-    if (openaiApiKeyRef) {
-      lines.push(`OPENAI_API_KEY_REF_B64=${encodeBase64(JSON.stringify(openaiApiKeyRef))}`);
-    }
-    if (telegramBotTokenRef) {
-      lines.push(`TELEGRAM_BOT_TOKEN_REF_B64=${encodeBase64(JSON.stringify(telegramBotTokenRef))}`);
-    }
-
-    const text = lines.join("\n") + "\n";
+    const text = buildEnvFileContent({
+      config,
+      inferenceProvider,
+      isVertex,
+      suggestedNamespace,
+      anthropicApiKeyRef,
+      openaiApiKeyRef,
+      telegramBotTokenRef,
+    });
     const blob = new Blob([text], { type: "text/plain;charset=utf-8" });
     const href = URL.createObjectURL(blob);
     const a = document.createElement("a");
@@ -1284,8 +721,10 @@ export default function DeployForm({ onDeployStarted }: Props) {
                   Enable Kagenti A2A
                 </label>
                 <div className="hint">
-                  Assumes a Kagenti stack is already running. The installer will add the Kagenti labels,
-                  namespace ConfigMaps, A2A bridge sidecar, built-in A2A skill, and the <code>AgentCard</code> resource.
+                  Assumes a Kagenti stack is already running. On OpenShift, this now assumes the cluster-admin
+                  prereq was installed with <code>setup-kagenti.sh</code>; the installer adds the Kagenti label,
+                  OpenClaw A2A bridge resources, and the <code>AgentCard</code> resource, while the Kagenti
+                  namespace controller handles the Kagenti ConfigMaps, RoleBindings, and SCC wiring.
                 </div>
               </div>
 
@@ -1296,7 +735,7 @@ export default function DeployForm({ onDeployStarted }: Props) {
                       <label>Keycloak Realm</label>
                       <input
                         type="text"
-                        placeholder="demo"
+                        placeholder="kagenti"
                         value={config.a2aRealm}
                         onChange={(e) => update("a2aRealm", e.target.value)}
                       />
@@ -1453,401 +892,21 @@ export default function DeployForm({ onDeployStarted }: Props) {
           </div>
         </div>
 
-        <h3 style={{ marginTop: "1.5rem" }}>Inference Provider</h3>
-
-        <div className="form-group">
-          <label>Primary Provider</label>
-          <select
-            value={inferenceProvider}
-            onChange={(e) => {
-              setInferenceProvider(e.target.value as InferenceProvider);
-              update("agentModel", ""); // Fix for #23: clear provider-specific model on switch
-            }}
-          >
-            {PROVIDER_OPTIONS.map((p) => (
-              <option key={p.id} value={p.id}>{p.label}</option>
-            ))}
-          </select>
-        <div className="hint">
-            {PROVIDER_OPTIONS.find((p) => p.id === inferenceProvider)?.desc}. This controls the default primary route for the deployment.
-          </div>
-        </div>
-
-        <div className="form-group" style={{ marginTop: "0.75rem" }}>
-          <label>Primary Model</label>
-          <input
-            type="text"
-            placeholder={
-              isVertex && config.litellmProxy
-                ? (inferenceProvider === "vertex-anthropic" ? "claude-sonnet-4-6" : "gemini-2.5-pro")
-                : (MODEL_DEFAULTS[inferenceProvider] || "model-id")
-            }
-            value={config.agentModel}
-            onChange={(e) => update("agentModel", e.target.value)}
-          />
-          <div className="hint">
-            {config.agentModel
-              ? "Custom primary model override"
-              : isVertex && config.litellmProxy
-                ? `Leave blank for default (routed through LiteLLM proxy). ${PROXY_MODEL_HINTS[inferenceProvider] || MODEL_HINTS[inferenceProvider]}`
-                : `Leave blank for default${MODEL_DEFAULTS[inferenceProvider] ? ` (${MODEL_DEFAULTS[inferenceProvider]})` : ""}. ${MODEL_HINTS[inferenceProvider]}`}
-          </div>
-        </div>
-
-        <div className="form-group" style={{ marginTop: "0.75rem" }}>
-          <label style={{ display: "flex", alignItems: "center", gap: "0.5rem" }}>
-            <input
-              type="checkbox"
-              checked={config.openaiCompatibleEndpointsEnabled}
-              onChange={(e) =>
-                setConfig((prev) => ({ ...prev, openaiCompatibleEndpointsEnabled: e.target.checked }))
-              }
-              style={{ width: "auto" }}
-            />
-            Enable OpenAI-compatible API endpoints
-          </label>
-          <div className="hint">
-            Exposes <code>/v1/chat/completions</code>, <code>/v1/responses</code>, and <code>/v1/models</code> for OpenAI-compatible clients. Disable this to remove those endpoints from the gateway.
-          </div>
-        </div>
-
-        <details style={{ marginTop: "0.75rem" }}>
-          <summary style={{ cursor: "pointer", fontWeight: 600 }}>
-            Additional Providers & Fallbacks
-            <span style={{ color: "var(--text-secondary)", fontWeight: "normal" }}>
-              {" "}Additional credentials and endpoint settings
-            </span>
-          </summary>
-
-          <div className="card" style={{ marginTop: "0.75rem" }}>
-            <div className="hint" style={{ marginBottom: "0.75rem" }}>
-              The selected primary provider and model above control the default route. The settings below are saved independently so Anthropic, OpenAI, and OpenAI-compatible endpoints can also be used for fallbacks.
-            </div>
-            <div className="hint" style={{ marginBottom: "0.75rem" }}>
-              Configure these independently when you want one deployment to have:
-              {" "}Anthropic (<code>ANTHROPIC_API_KEY</code>),
-              {" "}OpenAI (<code>OPENAI_API_KEY</code>), and
-              {" "}an OpenAI-compatible endpoint with its own token (<code>MODEL_ENDPOINT</code> + <code>MODEL_ENDPOINT_API_KEY</code>).
-            </div>
-            <div className="form-group">
-              <label>Anthropic API Key</label>
-              <input
-                type="password"
-                autoComplete="new-password"
-                placeholder={defaults?.hasAnthropicKey ? "(using key from environment)" : "sk-ant-..."}
-                value={config.anthropicApiKey}
-                onChange={(e) => update("anthropicApiKey", e.target.value)}
-              />
-              <div className="hint">
-                {defaults?.hasAnthropicKey
-                  ? "Detected ANTHROPIC_API_KEY from server environment — leave blank to use it"
-                  : "Saved for Anthropic primary or fallback usage."}
-              </div>
-            </div>
-            <div className="form-group">
-              <label>Anthropic Model</label>
-              <input
-                type="text"
-                placeholder="e.g., claude-sonnet-4-6"
-                value={config.anthropicModel}
-                onChange={(e) => update("anthropicModel", e.target.value)}
-              />
-              <div className="hint">
-                Adds this Anthropic model to the OpenClaw model picker as <code>anthropic/&lt;model&gt;</code>.
-              </div>
-            </div>
-
-            <div className="form-group">
-              <label>OpenAI API Key</label>
-              <input
-                type="password"
-                autoComplete="new-password"
-                placeholder={defaults?.hasOpenaiKey ? "(using key from environment)" : "sk-..."}
-                value={config.openaiApiKey}
-                onChange={(e) => update("openaiApiKey", e.target.value)}
-              />
-              <div className="hint">
-                {defaults?.hasOpenaiKey
-                  ? "Detected OPENAI_API_KEY from server environment — leave blank to use it"
-                  : "Saved for OpenAI primary or fallback usage only."}
-              </div>
-            </div>
-            <div className="form-group">
-              <label>OpenAI Model</label>
-              <input
-                type="text"
-                placeholder="e.g., gpt-5"
-                value={config.openaiModel}
-                onChange={(e) => update("openaiModel", e.target.value)}
-              />
-              <div className="hint">
-                Adds this OpenAI model to the OpenClaw model picker as <code>openai/&lt;model&gt;</code>.
-              </div>
-            </div>
-
-            {isVertex && (
-              <>
-                {inferenceProvider === "vertex-google"
-                  && gcpDefaults?.credentialType === "authorized_user"
-                  && !config.gcpServiceAccountJson && (
-                  <div style={{
-                    marginBottom: "1rem",
-                    padding: "0.5rem 0.75rem",
-                    background: "rgba(231, 76, 60, 0.1)",
-                    border: "1px solid rgba(231, 76, 60, 0.3)",
-                    borderRadius: "6px",
-                    fontSize: "0.85rem",
-                    color: "#e74c3c",
-                  }}>
-                    Your environment credentials are Application Default Credentials (from <code>gcloud auth</code>),
-                    which are not supported by Gemini on Vertex. Either upload a Service Account JSON below,
-                    or switch to Google Vertex AI (Claude) which works with Application Default Credentials.
-                  </div>
-                )}
-
-                <div className="form-row">
-                  <div className="form-group">
-                    <label>GCP Project ID</label>
-                    <input
-                      type="text"
-                      placeholder="my-gcp-project"
-                      value={config.googleCloudProject}
-                      onChange={(e) => update("googleCloudProject", e.target.value)}
-                    />
-                    {gcpDefaults?.sources.projectId && config.googleCloudProject === gcpDefaults.projectId ? (
-                      <div className="hint">from {gcpDefaults.sources.projectId}</div>
-                    ) : !config.googleCloudProject && (
-                      <div className="hint">Auto-extracted from credentials JSON if not set</div>
-                    )}
-                  </div>
-                  <div className="form-group">
-                    <label>GCP Region</label>
-                    <input
-                      type="text"
-                      placeholder={inferenceProvider === "vertex-anthropic" ? "us-east5 (default)" : "us-central1 (default)"}
-                      value={config.googleCloudLocation}
-                      onChange={(e) => update("googleCloudLocation", e.target.value)}
-                    />
-                    {gcpDefaults?.sources.location && config.googleCloudLocation === gcpDefaults.location ? (
-                      <div className="hint">from {gcpDefaults.sources.location}</div>
-                    ) : !config.googleCloudLocation && (
-                      <div className="hint">
-                        Defaults to {inferenceProvider === "vertex-anthropic" ? "us-east5" : "us-central1"} if not set
-                      </div>
-                    )}
-                  </div>
-                </div>
-
-                <div className="form-group">
-                  <label>Google Cloud Credentials (JSON)</label>
-                  <div style={{ display: "flex", gap: "0.5rem", alignItems: "center" }}>
-                    {config.gcpServiceAccountJson ? (
-                      <div
-                        style={{
-                          flex: 1,
-                          padding: "0.5rem 0.75rem",
-                          background: "var(--bg-primary)",
-                          border: "1px solid var(--border)",
-                          borderRadius: "6px",
-                          fontFamily: "monospace",
-                          fontSize: "0.85rem",
-                          color: "var(--text-secondary)",
-                        }}
-                      >
-                        {(() => {
-                          try {
-                            const parsed = JSON.parse(config.gcpServiceAccountJson);
-                            return `${parsed.client_email || "service account"} (${parsed.project_id || "unknown project"})`;
-                          } catch {
-                            return "credentials loaded";
-                          }
-                        })()}
-                      </div>
-                    ) : (
-                      <input
-                        type="text"
-                        placeholder={
-                          gcpDefaults?.hasServiceAccountJson
-                            ? `Using credentials from ${gcpDefaults.sources.credentials}`
-                            : "/path/to/service-account.json"
-                        }
-                        value={config.gcpServiceAccountPath}
-                        onChange={(e) => update("gcpServiceAccountPath", e.target.value)}
-                        style={{ flex: 1 }}
-                      />
-                    )}
-                    <label
-                      className="btn btn-ghost"
-                      style={{ cursor: "pointer", whiteSpace: "nowrap" }}
-                    >
-                      {config.gcpServiceAccountJson ? "Change" : "Browse"}
-                      <input
-                        type="file"
-                        accept=".json"
-                        style={{ display: "none" }}
-                        onChange={(e) => {
-                          const file = e.target.files?.[0];
-                          if (!file) return;
-                          const reader = new FileReader();
-                          reader.onload = () => {
-                            const text = reader.result as string;
-                            update("gcpServiceAccountJson", text);
-                            update("gcpServiceAccountPath", "");
-                            if (!config.googleCloudProject) {
-                              try {
-                                const parsed = JSON.parse(text);
-                                if (parsed.project_id) {
-                                  update("googleCloudProject", parsed.project_id);
-                                }
-                              } catch { /* ignore */ }
-                            }
-                          };
-                          reader.readAsText(file);
-                        }}
-                      />
-                    </label>
-                    {config.gcpServiceAccountJson && (
-                      <button
-                        className="btn btn-ghost"
-                        onClick={() => update("gcpServiceAccountJson", "")}
-                      >
-                        Clear
-                      </button>
-                    )}
-                  </div>
-                  <div className="hint">
-                    Type a path to a credentials JSON file, or use Browse to upload one.
-                    {gcpDefaults?.hasServiceAccountJson && !config.gcpServiceAccountJson && !config.gcpServiceAccountPath
-                      && " Leave blank to use credentials detected from environment."}
-                  </div>
-                </div>
-
-                <div className="form-group">
-                  <label style={{ display: "flex", alignItems: "center", gap: "0.5rem" }}>
-                    <input
-                      type="checkbox"
-                      checked={config.litellmProxy}
-                      onChange={(e) =>
-                        setConfig((prev) => ({ ...prev, litellmProxy: e.target.checked }))
-                      }
-                      style={{ width: "auto" }}
-                    />
-                    Use LiteLLM proxy (recommended)
-                  </label>
-                  <div className="hint">
-                    Runs a LiteLLM sidecar that handles Vertex AI authentication.
-                    GCP credentials stay in the proxy container and are never exposed to the agent.
-                    {!config.litellmProxy && (
-                      <span style={{ color: "#e67e22" }}>
-                        {" "}Disabled: credentials will be passed directly to the agent container.
-                      </span>
-                    )}
-                  </div>
-                  {config.litellmProxy && (
-                    <div style={{
-                      marginTop: "0.5rem",
-                      padding: "0.5rem 0.75rem",
-                      background: "rgba(52, 152, 219, 0.1)",
-                      border: "1px solid rgba(52, 152, 219, 0.3)",
-                      borderRadius: "6px",
-                      fontSize: "0.85rem",
-                      color: "var(--text-secondary)",
-                    }}>
-                      The first deployment will pull both the OpenClaw image and the LiteLLM proxy
-                      image (<code>ghcr.io/berriai/litellm:v1.82.3-stable.patch.2</code>, ~1.5 GB).
-                      This may take several minutes. You can pre-pull
-                      with: <code>{mode === "kubernetes" ? "crictl pull" : "podman pull"} ghcr.io/berriai/litellm:v1.82.3-stable.patch.2</code>
-                    </div>
-                  )}
-                </div>
-              </>
-            )}
-
-            <div className="form-group">
-              <label>OpenAI-Compatible Model Endpoint</label>
-              <div style={{ display: "flex", gap: "0.5rem", alignItems: "center" }}>
-                <input
-                  type="text"
-                  placeholder="http://vllm.openclaw-llms.svc.cluster.local/v1"
-                  value={config.modelEndpoint}
-                  onChange={(e) => update("modelEndpoint", e.target.value)}
-                  style={{ flex: 1 }}
-                />
-                <button
-                  type="button"
-                  className="btn btn-ghost"
-                  onClick={fetchModelEndpointOptions}
-                  disabled={loadingModelEndpointOptions}
-                >
-                  {loadingModelEndpointOptions ? "Fetching..." : "Fetch Models"}
-                </button>
-              </div>
-              <div className="hint">
-                Optional. Save a local or open-source OpenAI-compatible endpoint here for primary use or fallback routing.
-              </div>
-              <div className="hint" style={{ marginTop: "0.35rem" }}>
-                If you paste just the service URL, the installer will normalize it to a <code>/v1</code> API base for runtime requests.
-              </div>
-              {modelEndpointOptionsError && (
-                <div className="hint" style={{ color: "#e74c3c", marginTop: "0.35rem" }}>
-                  {modelEndpointOptionsError}
-                </div>
-              )}
-            </div>
-            <div className="form-group">
-              <label>OpenAI-Compatible Model Name</label>
-              {modelEndpointOptions.length > 0 ? (
-                <select
-                  value={config.modelEndpointModel}
-                  onChange={(e) => {
-                    const selected = modelEndpointOptions.find((option) => option.id === e.target.value);
-                    setConfig((prev) => ({
-                      ...prev,
-                      modelEndpointModel: e.target.value,
-                      modelEndpointModelLabel: selected?.name || e.target.value,
-                    }));
-                  }}
-                >
-                  {modelEndpointOptions.map((option) => (
-                    <option key={option.id} value={option.id}>
-                      {option.name === option.id ? option.id : `${option.name} (${option.id})`}
-                    </option>
-                  ))}
-                </select>
-              ) : (
-                <input
-                  type="text"
-                  placeholder="e.g., mistral-small-24b-w8a8"
-                  value={config.modelEndpointModel}
-                  onChange={(e) =>
-                    setConfig((prev) => ({
-                      ...prev,
-                      modelEndpointModel: e.target.value,
-                      modelEndpointModelLabel: e.target.value,
-                    }))
-                  }
-                />
-              )}
-              <div className="hint">
-                The model ID served by that endpoint. Use the exact name the endpoint expects, such as <code>mistral-small-24b-w8a8</code>.
-              </div>
-            </div>
-            <div className="form-group">
-              <label>OpenAI-Compatible Endpoint API Key (`MODEL_ENDPOINT_API_KEY`)</label>
-              <input
-                type="password"
-                autoComplete="new-password"
-                placeholder="API key for the OpenAI-compatible endpoint"
-                value={config.modelEndpointApiKey}
-                onChange={(e) => update("modelEndpointApiKey", e.target.value)}
-              />
-              <div className="hint">
-                Separate from <code>OPENAI_API_KEY</code>. Use this when your OpenAI-compatible endpoint requires its own token.
-              </div>
-            </div>
-
-          </div>
-        </details>
+        <ProviderSection
+          config={config}
+          defaults={defaults}
+          fetchModelEndpointOptions={fetchModelEndpointOptions}
+          gcpDefaults={gcpDefaults}
+          inferenceProvider={inferenceProvider}
+          isVertex={isVertex}
+          loadingModelEndpointOptions={loadingModelEndpointOptions}
+          mode={mode}
+          modelEndpointOptions={modelEndpointOptions}
+          modelEndpointOptionsError={modelEndpointOptionsError}
+          setConfig={setConfig}
+          setInferenceProvider={setInferenceProvider}
+          update={update}
+        />
 
         <h3 style={{ marginTop: "1.5rem" }}>Observability</h3>
 
@@ -1973,397 +1032,16 @@ export default function DeployForm({ onDeployStarted }: Props) {
           </>
         )}
 
-        <h3 style={{ marginTop: "1.5rem" }}>Sandbox</h3>
+        <SandboxSection
+          config={config}
+          update={update}
+          setConfig={setConfig}
+        />
 
-        <div className="form-group">
-          <label style={{ display: "flex", gap: "0.5rem", alignItems: "center" }}>
-            <input
-              type="checkbox"
-              checked={config.sandboxEnabled}
-              onChange={(e) =>
-                setConfig((prev) => ({ ...prev, sandboxEnabled: e.target.checked }))
-              }
-            />
-            Enable SSH sandbox backend
-          </label>
-          <div className="hint">
-            Recommended path for this installer on both local containers and Kubernetes.
-          </div>
-        </div>
-
-        {config.sandboxEnabled && (
-          <>
-            <div className="form-row">
-              <div className="form-group">
-                <label>Sandbox Mode</label>
-                <select
-                  value={config.sandboxMode}
-                  onChange={(e) => update("sandboxMode", e.target.value)}
-                >
-                  <option value="all">all</option>
-                  <option value="non-main">non-main</option>
-                  <option value="off">off</option>
-                </select>
-              </div>
-              <div className="form-group">
-                <label>Sandbox Scope</label>
-                <select
-                  value={config.sandboxScope}
-                  onChange={(e) => update("sandboxScope", e.target.value)}
-                >
-                  <option value="session">session</option>
-                  <option value="agent">agent</option>
-                  <option value="shared">shared</option>
-                </select>
-              </div>
-            </div>
-
-            <div className="form-row">
-              <div className="form-group">
-                <label>Workspace Access</label>
-                <select
-                  value={config.sandboxWorkspaceAccess}
-                  onChange={(e) => update("sandboxWorkspaceAccess", e.target.value)}
-                >
-                  <option value="rw">rw</option>
-                  <option value="ro">ro</option>
-                  <option value="none">none</option>
-                </select>
-              </div>
-              <div className="form-group">
-                <label>Remote Workspace Root</label>
-                <input
-                  type="text"
-                  placeholder="/tmp/openclaw-sandboxes"
-                  value={config.sandboxSshWorkspaceRoot}
-                  onChange={(e) => update("sandboxSshWorkspaceRoot", e.target.value)}
-                />
-              </div>
-            </div>
-
-            <div className="form-group">
-              <label style={{ display: "flex", gap: "0.5rem", alignItems: "center" }}>
-                <input
-                  type="checkbox"
-                  checked={config.sandboxToolPolicyEnabled}
-                  onChange={(e) =>
-                    setConfig((prev) => ({ ...prev, sandboxToolPolicyEnabled: e.target.checked }))
-                  }
-                />
-                Customize sandbox tool baseline
-              </label>
-              <div className="hint">
-                Optional persistent baseline for sandboxed tools. This is intentionally much smaller than the full gateway UI.
-              </div>
-            </div>
-
-            {config.sandboxToolPolicyEnabled && (
-              <div className="form-row" style={{ flexWrap: "wrap", gap: "1rem 1.5rem", marginBottom: "1rem" }}>
-                <label style={{ display: "flex", gap: "0.5rem", alignItems: "center" }}>
-                  <input
-                    type="checkbox"
-                    checked={config.sandboxToolAllowFiles}
-                    onChange={(e) =>
-                      setConfig((prev) => ({ ...prev, sandboxToolAllowFiles: e.target.checked }))
-                    }
-                  />
-                  File tools
-                </label>
-                <label style={{ display: "flex", gap: "0.5rem", alignItems: "center" }}>
-                  <input
-                    type="checkbox"
-                    checked={config.sandboxToolAllowSessions}
-                    onChange={(e) =>
-                      setConfig((prev) => ({ ...prev, sandboxToolAllowSessions: e.target.checked }))
-                    }
-                  />
-                  Session tools
-                </label>
-                <label style={{ display: "flex", gap: "0.5rem", alignItems: "center" }}>
-                  <input
-                    type="checkbox"
-                    checked={config.sandboxToolAllowMemory}
-                    onChange={(e) =>
-                      setConfig((prev) => ({ ...prev, sandboxToolAllowMemory: e.target.checked }))
-                    }
-                  />
-                  Memory tools
-                </label>
-                <label style={{ display: "flex", gap: "0.5rem", alignItems: "center" }}>
-                  <input
-                    type="checkbox"
-                    checked={config.sandboxToolAllowRuntime}
-                    onChange={(e) =>
-                      setConfig((prev) => ({ ...prev, sandboxToolAllowRuntime: e.target.checked }))
-                    }
-                  />
-                  Runtime tools (`exec`, `bash`, `process`)
-                </label>
-                <label style={{ display: "flex", gap: "0.5rem", alignItems: "center" }}>
-                  <input
-                    type="checkbox"
-                    checked={config.sandboxToolAllowBrowser}
-                    onChange={(e) =>
-                      setConfig((prev) => ({ ...prev, sandboxToolAllowBrowser: e.target.checked }))
-                    }
-                  />
-                  Browser and canvas
-                </label>
-                <label style={{ display: "flex", gap: "0.5rem", alignItems: "center" }}>
-                  <input
-                    type="checkbox"
-                    checked={config.sandboxToolAllowAutomation}
-                    onChange={(e) =>
-                      setConfig((prev) => ({ ...prev, sandboxToolAllowAutomation: e.target.checked }))
-                    }
-                  />
-                  Automation tools
-                </label>
-                <label style={{ display: "flex", gap: "0.5rem", alignItems: "center" }}>
-                  <input
-                    type="checkbox"
-                    checked={config.sandboxToolAllowMessaging}
-                    onChange={(e) =>
-                      setConfig((prev) => ({ ...prev, sandboxToolAllowMessaging: e.target.checked }))
-                    }
-                  />
-                  Messaging tools
-                </label>
-              </div>
-            )}
-
-            <div className="form-group">
-              <label>SSH Target</label>
-              <input
-                type="text"
-                placeholder="user@gateway-host:22"
-                value={config.sandboxSshTarget}
-                onChange={(e) => update("sandboxSshTarget", e.target.value)}
-              />
-              <div className="hint">
-                Required. OpenClaw will run sandboxed tools on this remote host.
-              </div>
-            </div>
-
-            <div className="form-row">
-              <label style={{ display: "flex", gap: "0.5rem", alignItems: "center" }}>
-                <input
-                  type="checkbox"
-                  checked={config.sandboxSshStrictHostKeyChecking}
-                  onChange={(e) =>
-                    setConfig((prev) => ({
-                      ...prev,
-                      sandboxSshStrictHostKeyChecking: e.target.checked,
-                    }))}
-                />
-                Strict host key checking
-              </label>
-              <label style={{ display: "flex", gap: "0.5rem", alignItems: "center" }}>
-                <input
-                  type="checkbox"
-                  checked={config.sandboxSshUpdateHostKeys}
-                  onChange={(e) =>
-                    setConfig((prev) => ({
-                      ...prev,
-                      sandboxSshUpdateHostKeys: e.target.checked,
-                    }))}
-                />
-                Update host keys
-              </label>
-            </div>
-
-            <div className="form-group">
-              <label>SSH Private Key</label>
-              <input
-                type="text"
-                placeholder="/path/to/id_ed25519"
-                value={config.sandboxSshIdentityPath}
-                onChange={(e) => update("sandboxSshIdentityPath", e.target.value)}
-              />
-              <div className="hint">Path on the installer host to the private key file.</div>
-            </div>
-
-            <div className="form-group">
-              <label>
-                SSH Certificate
-                <span style={{ color: "var(--text-secondary)", fontWeight: "normal" }}>
-                  {" "}(optional)
-                </span>
-              </label>
-              <input
-                type="text"
-                placeholder="/path/to/id_ed25519-cert.pub"
-                value={config.sandboxSshCertificatePath}
-                onChange={(e) => update("sandboxSshCertificatePath", e.target.value)}
-                style={{ marginBottom: "0.5rem" }}
-              />
-              <textarea
-                rows={4}
-                placeholder="ssh-ed25519-cert-v01@openssh.com ..."
-                value={config.sandboxSshCertificate}
-                onChange={(e) => update("sandboxSshCertificate", e.target.value)}
-              />
-              <div className="hint">Type a path on the installer host, or paste the certificate directly.</div>
-            </div>
-
-            <div className="form-group">
-              <label>
-                Known Hosts
-                <span style={{ color: "var(--text-secondary)", fontWeight: "normal" }}>
-                  {" "}(optional)
-                </span>
-              </label>
-              <input
-                type="text"
-                placeholder="/path/to/known_hosts"
-                value={config.sandboxSshKnownHostsPath}
-                onChange={(e) => update("sandboxSshKnownHostsPath", e.target.value)}
-                style={{ marginBottom: "0.5rem" }}
-              />
-              <textarea
-                rows={4}
-                placeholder="gateway-host ssh-ed25519 AAAA..."
-                value={config.sandboxSshKnownHosts}
-                onChange={(e) => update("sandboxSshKnownHosts", e.target.value)}
-              />
-              <div className="hint">Type a path on the installer host, or paste known_hosts entries directly.</div>
-            </div>
-          </>
-        )}
-
-        <details style={{ marginTop: "1.5rem" }}>
-          <summary style={{ cursor: "pointer", fontWeight: 600 }}>Advanced: Experimental External Secret Providers</summary>
-          <div className="card" style={{ marginTop: "0.75rem" }}>
-            <div className="hint" style={{ marginBottom: "0.75rem" }}>
-              Only use this if your secrets come from an external provider such as Vault, a mounted file,
-              or a custom command. Most users should leave this closed and just enter credentials in the normal fields above.
-            </div>
-            <div className="form-group">
-              <label>Secret Providers JSON (optional)</label>
-              <textarea
-                rows={6}
-                placeholder={`{\n  "default": { "source": "env" },\n  "vault_openai": {\n    "source": "exec",\n    "command": "/usr/local/bin/vault",\n    "args": ["kv", "get", "-field=OPENAI_API_KEY", "secret/openclaw"],\n    "passEnv": ["VAULT_ADDR", "VAULT_TOKEN"]\n  }\n}`}
-                value={config.secretsProvidersJson}
-                onChange={(e) => update("secretsProvidersJson", e.target.value)}
-              />
-              <div className="hint">
-                Optional <code>secrets.providers</code> object. Runtime prerequisites still need to exist
-                inside the OpenClaw environment.
-              </div>
-            </div>
-
-            <div className="card" style={{ marginBottom: "1rem" }}>
-              <div className="form-row">
-                <div className="form-group">
-                  <label>Anthropic SecretRef Source</label>
-                  <select
-                    value={config.anthropicApiKeyRefSource}
-                    onChange={(e) => update("anthropicApiKeyRefSource", e.target.value)}
-                  >
-                    <option value="env">env</option>
-                    <option value="file">file</option>
-                    <option value="exec">exec</option>
-                  </select>
-                </div>
-                <div className="form-group">
-                  <label>Anthropic SecretRef Provider</label>
-                  <input
-                    type="text"
-                    placeholder="default"
-                    value={config.anthropicApiKeyRefProvider}
-                    onChange={(e) => update("anthropicApiKeyRefProvider", e.target.value)}
-                  />
-                </div>
-              </div>
-              <div className="form-group">
-                <label>Anthropic SecretRef ID</label>
-                <input
-                  type="text"
-                  placeholder="ANTHROPIC_API_KEY or /providers/anthropic/apiKey or providers/anthropic/apiKey"
-                  value={config.anthropicApiKeyRefId}
-                  onChange={(e) => update("anthropicApiKeyRefId", e.target.value)}
-                />
-                <div className="hint">
-                  Optional override. Leave blank to use the installer-managed env-backed SecretRef automatically.
-                </div>
-              </div>
-            </div>
-
-            <div className="card" style={{ marginBottom: "1rem" }}>
-              <div className="form-row">
-                <div className="form-group">
-                  <label>OpenAI SecretRef Source</label>
-                  <select
-                    value={config.openaiApiKeyRefSource}
-                    onChange={(e) => update("openaiApiKeyRefSource", e.target.value)}
-                  >
-                    <option value="env">env</option>
-                    <option value="file">file</option>
-                    <option value="exec">exec</option>
-                  </select>
-                </div>
-                <div className="form-group">
-                  <label>OpenAI SecretRef Provider</label>
-                  <input
-                    type="text"
-                    placeholder="default"
-                    value={config.openaiApiKeyRefProvider}
-                    onChange={(e) => update("openaiApiKeyRefProvider", e.target.value)}
-                  />
-                </div>
-              </div>
-              <div className="form-group">
-                <label>OpenAI SecretRef ID</label>
-                <input
-                  type="text"
-                  placeholder="OPENAI_API_KEY or /providers/openai/apiKey or providers/openai/apiKey"
-                  value={config.openaiApiKeyRefId}
-                  onChange={(e) => update("openaiApiKeyRefId", e.target.value)}
-                />
-                <div className="hint">
-                  Optional override. Leave blank to use the installer-managed env-backed SecretRef automatically.
-                </div>
-              </div>
-            </div>
-
-            <div className="card" style={{ marginBottom: "1rem" }}>
-              <div className="form-row">
-                <div className="form-group">
-                  <label>Telegram SecretRef Source</label>
-                  <select
-                    value={config.telegramBotTokenRefSource}
-                    onChange={(e) => update("telegramBotTokenRefSource", e.target.value)}
-                  >
-                    <option value="env">env</option>
-                    <option value="file">file</option>
-                    <option value="exec">exec</option>
-                  </select>
-                </div>
-                <div className="form-group">
-                  <label>Telegram SecretRef Provider</label>
-                  <input
-                    type="text"
-                    placeholder="default"
-                    value={config.telegramBotTokenRefProvider}
-                    onChange={(e) => update("telegramBotTokenRefProvider", e.target.value)}
-                  />
-                </div>
-              </div>
-              <div className="form-group">
-                <label>Telegram SecretRef ID</label>
-                <input
-                  type="text"
-                  placeholder="TELEGRAM_BOT_TOKEN or /channels/telegram/botToken or channels/telegram/botToken"
-                  value={config.telegramBotTokenRefId}
-                  onChange={(e) => update("telegramBotTokenRefId", e.target.value)}
-                />
-                <div className="hint">
-                  Optional override. Leave blank to use the installer-managed env-backed SecretRef automatically.
-                </div>
-              </div>
-            </div>
-          </div>
-        </details>
+        <SecretProvidersSection
+          config={config}
+          update={update}
+        />
 
         <div style={{ marginTop: "1.5rem" }}>
           {!isValid && (
